+++ client/src/Components/Authentication/MyProfile/MyProfile.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
+import { useOutletContext } from "react-router-dom";
 import Head from "../../Common/Header/Head";
 import Footer from "../../Common/FooterView/Footer";
 import TabsView from "../TabsView/TabsView";
@@ -20,26 +21,27 @@ const MyProfile = () => {
       }
     });
 
-  return (
-    !user ? loginWithRedirect() :
-    isAuthenticated && (
-      <div>
-        <Head />        
-        <div>
-          <h2>Hola, {user.nickname}.</h2>
-          {/* <p>{user.email}</p> */}
+  const [localUserData] = useOutletContext();
+  // console.log('habra llegado herencia?? ', localUserData);
 
-        </div>
-        <aside>
-          <TabsView />
-        </aside>
-          <button onClick={() => logoutWithRedirect()}>
-            Cerrar sesión
-          </button>
-          <TakeUserData />
-        <Footer />
+  return (
+    // !user ? loginWithRedirect() :
+    // isAuthenticated && (
+    <div>
+      <Head />
+      <div>
+        <h2>Hola, {user.nickname}.</h2>
       </div>
-    )
+      <aside>
+        <TabsView localUserData={localUserData} />
+      </aside>
+      <button onClick={() => logoutWithRedirect()}>
+        Cerrar sesión
+      </button>
+      <TakeUserData />
+      <Footer />
+    </div>
+    //)
   )
 };
