+++ client/src/Components/Common/ProductCard/ProductCard.jsx
@@ -21,7 +21,7 @@ const ProductCard = ({product}) => {
   }
 
 
-  console.log(product)
+  // console.log(product)
 
   return (
